+++ Sources/Contour/Sending/AbstractSender.cs
@@ -38,11 +38,17 @@ namespace Contour.Sending
         // TODO: refactor, don't copy filters
 
         /// <summary>
-        /// Инициализирует новый экземпляр класса <see cref="AbstractSender"/>.
+        /// Initializes a new instance of the <see cref="AbstractSender"/> class. 
         /// </summary>
-        /// <param name="endpoint">Конечная точка, от имени которой работает отправитель.</param>
-        /// <param name="configuration">Конфигурация отправителя.</param>
-        /// <param name="filters">Список фильтров обработки сообщения.</param>
+        /// <param name="endpoint">
+        /// Sender's endpoint
+        /// </param>
+        /// <param name="configuration">
+        /// Sender's configuration
+        /// </param>
+        /// <param name="filters">
+        /// A list of message handling filters
+        /// </param>
         protected AbstractSender(IEndpoint endpoint, ISenderConfiguration configuration, IEnumerable<IMessageExchangeFilter> filters)
         {
             this.endpoint = endpoint;
@@ -88,12 +94,14 @@ namespace Contour.Sending
         public abstract void Dispose();
 
         /// <summary>
-        /// Отправляет сообщение в формате запрос-ответ.
+        /// Sends message using request-reply pattern. 
+        /// <see cref="Headers.CorrelationId"/> header from <see cref="headers"/> parameter is used to correlate the request with the reply, 
+        /// new one is generated if none is supplied.
         /// </summary>
-        /// <param name="payload">Сообщение запроса.</param>
-        /// <param name="headers">Заголовки запроса.</param>
-        /// <typeparam name="T">Тип сообщения ответа.</typeparam>
-        /// <returns>Задача выполнения запроса.</returns>
+        /// <param name="payload">Message payload.</param>
+        /// <param name="headers">Message headers.</param>
+        /// <typeparam name="T">Type of response.</typeparam>
+        /// <returns>Request processing task.</returns>
         public Task<T> Request<T>(object payload, IDictionary<string, object> headers) where T : class
         {
             var message = new Message(this.Configuration.Label, headers, payload);
@@ -111,24 +119,23 @@ namespace Contour.Sending
         }
 
         /// <summary>
-        /// Отправляет сообщение в формате запрос-ответ.
+        /// Sends message using request-reply pattern.
+        /// Copies all allowed message headers and generates new <see cref="Headers.CorrelationId"/> header. 
+        /// <seealso ref="https://github.com/SDVentures/Contour#contour-message-headers"/>.
         /// </summary>
-        /// <param name="payload">Сообщение запроса.</param>
-        /// <param name="options">Параметры запроса.</param>
-        /// <typeparam name="T">Тип сообщения ответа.</typeparam>
-        /// <returns>Задача выполнения запроса.</returns>
+        /// <param name="payload">Message payload.</param>
+        /// <param name="options">Request options.</param>
+        /// <typeparam name="T">Type of response.</typeparam>
+        /// <returns>Request processing task.</returns>
         public Task<T> Request<T>(object payload, RequestOptions options) where T : class
         {
             var headers = this.ApplyOptions(options);
-            if (!headers.ContainsKey(Headers.CorrelationId))
-            {
-                headers[Headers.CorrelationId] = Guid.NewGuid().ToString("n");
-            }
+            headers[Headers.CorrelationId] = Guid.NewGuid().ToString("n");
 
             return this.Request<T>(payload, headers);
         }
 
-            /// <summary>
+        /// <summary>
         /// Отправляет одностороннее сообщение.
         /// </summary>
         /// <param name="payload">Тело сообщения.</param>
@@ -155,28 +162,33 @@ namespace Contour.Sending
         }
 
         /// <summary>
-        /// Отправляет сообщение в формате запрос-ответ.
+        /// Sends message using request-reply pattern.
+        /// Copies all allowed message headers. And generates new <see cref="Headers.CorrelationId"/> header. 
+        /// <seealso ref="https://github.com/SDVentures/Contour#contour-message-headers"/>.
         /// </summary>
-        /// <param name="label">Метка отправляемого запроса.</param>
-        /// <param name="payload">Сообщение запроса.</param>
-        /// <param name="options">Параметры запроса.</param>
-        /// <typeparam name="T">Тип сообщения ответа.</typeparam>
-        /// <returns>Задача выполнения запроса.</returns>
+        /// <param name="label">Message label.</param>
+        /// <param name="payload">Message payload.</param>
+        /// <param name="options">Request options.</param>
+        /// <typeparam name="T">Type of response.</typeparam>
+        /// <returns>Request processing task.</returns>
         public Task<T> Request<T>(MessageLabel label, object payload, RequestOptions options) where T : class
         {
             var headers = this.ApplyOptions(options);
+            headers[Headers.CorrelationId] = Guid.NewGuid().ToString("n");
 
             return this.Request<T>(label, payload, headers);
         }
 
         /// <summary>
-        /// Отправляет сообщение в формате запрос-ответ.
+        /// Sends message using request-reply pattern. 
+        /// <see cref="Headers.CorrelationId"/> header from <see cref="headers"/> parameter is used to correlate the request with the reply, 
+        /// new one is generated if none is supplied.
         /// </summary>
-        /// <param name="label">Метка отправляемого запроса.</param>
-        /// <param name="payload">Сообщение запроса.</param>
-        /// <param name="headers">Заголовки запроса.</param>
-        /// <typeparam name="T">Тип сообщения ответа.</typeparam>
-        /// <returns>Задача выполнения запроса.</returns>
+        /// <param name="label">Message label.</param>
+        /// <param name="payload">Message payload.</param>
+        /// <param name="headers">Message headers.</param>
+        /// <typeparam name="T">Type of response.</typeparam>
+        /// <returns>Request processing task.</returns>
         public Task<T> Request<T>(MessageLabel label, object payload, IDictionary<string, object> headers) where T : class
         {
             if (!headers.ContainsKey(Headers.CorrelationId))
@@ -184,6 +196,7 @@ namespace Contour.Sending
                 headers[Headers.CorrelationId] = Guid.NewGuid().ToString("n");
             }
 
+            Logger.Trace(m => m("Invoking request to label [{0}] with payload [{1}] and correlationId = [{2}]", label, payload, headers[Headers.CorrelationId]));
             var message = new Message(this.Configuration.Label.Equals(MessageLabel.Any) ? label : this.Configuration.Label, headers, payload);
 
             var exchange = new MessageExchange(message, typeof(T));